pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinela"
version = "0.1.0"
description = "Coletor modular de notícias em portais com MongoDB"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Sentinela"}]
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pymongo>=4.6",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sse-starlette>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
sentinela-api = "sentinela.api:run"
sentinela-cli = "sentinela.cli:main"
